+++ lib/iris/tests/unit/fileformats/abf/test_ABFField.py
@@ -5,8 +5,8 @@
 # licensing details.
 """Unit tests for the `iris.fileformats.abf.ABFField` class."""
 
-from __future__ import (absolute_import, division, print_function)
-from six.moves import (filter, input, map, range, zip)  # noqa
+from __future__ import absolute_import, division, print_function
+from six.moves import filter, input, map, range, zip  # noqa
 
 # Import iris.tests first so that some things can be initialised before
 # importing anything else.
@@ -39,15 +39,15 @@ class MethodCounter(object):
 
 class Test_data(tests.IrisTest):
     def test_single_read(self):
-        path = '0000000000000000jan00000'
+        path = "0000000000000000jan00000"
         field = ABFField(path)
 
-        with mock.patch('iris.fileformats.abf.np.fromfile') as fromfile:
-            with MethodCounter('__getattr__') as getattr:
-                with MethodCounter('_read') as read:
+        with mock.patch("iris.fileformats.abf.np.fromfile") as fromfile:
+            with MethodCounter("__getattr__") as getattr:
+                with MethodCounter("_read") as read:
                     field.data
 
-        fromfile.assert_called_once_with(path, dtype='>u1')
+        fromfile.assert_called_once_with(path, dtype=">u1")
         self.assertEqual(getattr.count, 1)
         self.assertEqual(read.count, 1)
